remove bff db package
