+++ pkg/bff/api/v1/api.go
@@ -3,7 +3,7 @@ package api
 import (
 	"context"
 
-	"github.com/trisacrypto/directory/pkg/bff/db/models/v1"
+	"github.com/trisacrypto/directory/pkg/bff/models/v1"
 	members "github.com/trisacrypto/directory/pkg/gds/members/v1alpha1"
 )
