+++ pkg/bff/auth/clive/bindata.go
@@ -85,7 +85,7 @@ func faviconPng() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "favicon.png", size: 1693, mode: os.FileMode(0644), modTime: time.Unix(1659970605, 0)}
+	info := bindataFileInfo{name: "favicon.png", size: 1693, mode: os.FileMode(0644), modTime: time.Unix(1663788021, 0)}
 	a := &asset{bytes: bytes, info: info, digest: [32]uint8{0x38, 0x4a, 0x3c, 0xe2, 0x15, 0xfa, 0xea, 0x42, 0x57, 0xdc, 0xbf, 0xf6, 0x1a, 0x1f, 0x6c, 0x67, 0xef, 0xf9, 0x35, 0xb9, 0xfc, 0x1b, 0xaf, 0xe6, 0x30, 0xc, 0x86, 0x72, 0x45, 0xf2, 0x6c, 0x4e}}
 	return a, nil
 }
@@ -105,7 +105,7 @@ func indexHtml() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "index.html", size: 7101, mode: os.FileMode(0644), modTime: time.Unix(1659970605, 0)}
+	info := bindataFileInfo{name: "index.html", size: 7101, mode: os.FileMode(0644), modTime: time.Unix(1663788021, 0)}
 	a := &asset{bytes: bytes, info: info, digest: [32]uint8{0xf4, 0x44, 0x7d, 0xa7, 0x8a, 0xb5, 0xae, 0x30, 0x54, 0x50, 0x75, 0x4a, 0x62, 0xe0, 0xce, 0xad, 0xc5, 0xb8, 0x34, 0xf5, 0x5a, 0xe5, 0x2e, 0x68, 0x15, 0x44, 0xec, 0xb0, 0x5e, 0x5c, 0x39, 0xe8}}
 	return a, nil
 }
