+++ pkg/bff/errors.go
@@ -1,4 +1,4 @@
-package db
+package bff
 
 import "errors"
