+++ pkg/bff/models/v1/testdata/default_registration_form.json
