+++ pkg/bff/testdata/announcements.json
