+++ pkg/gds/emails/bindata.go
@@ -99,7 +99,7 @@ func deliver_certsHtml() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "deliver_certs.html", size: 1591, mode: os.FileMode(0644), modTime: time.Unix(1659970605, 0)}
+	info := bindataFileInfo{name: "deliver_certs.html", size: 1591, mode: os.FileMode(0644), modTime: time.Unix(1663788021, 0)}
 	a := &asset{bytes: bytes, info: info, digest: [32]uint8{0xd0, 0x1e, 0x2, 0xe1, 0xbe, 0xdf, 0x98, 0xd1, 0x19, 0x6e, 0x46, 0xeb, 0xf2, 0x2c, 0xef, 0x58, 0x78, 0x47, 0x12, 0xc4, 0x43, 0x70, 0x47, 0x70, 0x57, 0xcd, 0x97, 0xca, 0xf, 0x37, 0xa0, 0xa0}}
 	return a, nil
 }
@@ -119,7 +119,7 @@ func deliver_certsTxt() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "deliver_certs.txt", size: 1274, mode: os.FileMode(0644), modTime: time.Unix(1659970605, 0)}
+	info := bindataFileInfo{name: "deliver_certs.txt", size: 1274, mode: os.FileMode(0644), modTime: time.Unix(1663788021, 0)}
 	a := &asset{bytes: bytes, info: info, digest: [32]uint8{0x61, 0x39, 0xab, 0xc4, 0xc3, 0xc8, 0xb4, 0x7c, 0x13, 0xa, 0x1, 0x4c, 0xdb, 0xda, 0xec, 0x53, 0xb0, 0xb0, 0xc5, 0x15, 0x68, 0xce, 0xf1, 0xa3, 0x41, 0x22, 0x3d, 0x7, 0x33, 0x5, 0x1c, 0xcd}}
 	return a, nil
 }
@@ -139,7 +139,7 @@ func expires_admin_notificationHtml() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "expires_admin_notification.html", size: 1110, mode: os.FileMode(0644), modTime: time.Unix(1659970605, 0)}
+	info := bindataFileInfo{name: "expires_admin_notification.html", size: 1110, mode: os.FileMode(0644), modTime: time.Unix(1663788021, 0)}
 	a := &asset{bytes: bytes, info: info, digest: [32]uint8{0x6e, 0xc9, 0xe4, 0xcd, 0x40, 0x52, 0x48, 0x1f, 0x36, 0xc5, 0xc8, 0x90, 0xb8, 0xd, 0x2d, 0xd6, 0x9, 0xe5, 0xea, 0xd3, 0x66, 0xc5, 0xdc, 0x1f, 0xcf, 0x94, 0x9c, 0x6e, 0x45, 0x32, 0x17, 0xb4}}
 	return a, nil
 }
@@ -159,7 +159,7 @@ func expires_admin_notificationTxt() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "expires_admin_notification.txt", size: 816, mode: os.FileMode(0644), modTime: time.Unix(1659970605, 0)}
+	info := bindataFileInfo{name: "expires_admin_notification.txt", size: 816, mode: os.FileMode(0644), modTime: time.Unix(1663788021, 0)}
 	a := &asset{bytes: bytes, info: info, digest: [32]uint8{0x79, 0x27, 0x88, 0xda, 0x4, 0x2f, 0x83, 0x3, 0x99, 0xb0, 0x95, 0x27, 0x66, 0x77, 0x31, 0xcc, 0x28, 0x94, 0x88, 0x1a, 0x28, 0x83, 0xdc, 0x8e, 0xf3, 0x4c, 0xde, 0xec, 0xe6, 0x2c, 0xd, 0x83}}
 	return a, nil
 }
@@ -179,7 +179,7 @@ func reissuance_admin_notificationHtml() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "reissuance_admin_notification.html", size: 852, mode: os.FileMode(0644), modTime: time.Unix(1662834707, 0)}
+	info := bindataFileInfo{name: "reissuance_admin_notification.html", size: 852, mode: os.FileMode(0644), modTime: time.Unix(1663788021, 0)}
 	a := &asset{bytes: bytes, info: info, digest: [32]uint8{0xb1, 0x70, 0xee, 0xc6, 0x33, 0xf4, 0x59, 0x88, 0xf0, 0x23, 0xd0, 0xeb, 0x55, 0x2d, 0x97, 0x41, 0xb9, 0xad, 0x3d, 0xcc, 0xc4, 0xd6, 0xb4, 0x6a, 0x49, 0xab, 0x63, 0x3f, 0xba, 0x32, 0x44, 0xe6}}
 	return a, nil
 }
@@ -199,7 +199,7 @@ func reissuance_admin_notificationTxt() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "reissuance_admin_notification.txt", size: 567, mode: os.FileMode(0644), modTime: time.Unix(1662834707, 0)}
+	info := bindataFileInfo{name: "reissuance_admin_notification.txt", size: 567, mode: os.FileMode(0644), modTime: time.Unix(1663788021, 0)}
 	a := &asset{bytes: bytes, info: info, digest: [32]uint8{0x86, 0x8d, 0xef, 0xd0, 0x7b, 0xbb, 0x81, 0xe6, 0x53, 0xc0, 0xac, 0x0, 0xf1, 0xff, 0x11, 0x20, 0x11, 0x78, 0x5f, 0xe7, 0xcf, 0xbd, 0x66, 0xa4, 0x98, 0xe5, 0xcf, 0x55, 0x6f, 0x53, 0x11, 0x0}}
 	return a, nil
 }
@@ -219,7 +219,7 @@ func reissuance_reminderHtml() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "reissuance_reminder.html", size: 2042, mode: os.FileMode(0644), modTime: time.Unix(1659970605, 0)}
+	info := bindataFileInfo{name: "reissuance_reminder.html", size: 2042, mode: os.FileMode(0644), modTime: time.Unix(1663788021, 0)}
 	a := &asset{bytes: bytes, info: info, digest: [32]uint8{0x92, 0x71, 0x2a, 0x5e, 0xc3, 0x5a, 0x2f, 0xf4, 0xb8, 0x18, 0x15, 0xe3, 0x23, 0x40, 0x85, 0xef, 0xfe, 0x69, 0x3, 0xaf, 0xf1, 0xe4, 0x72, 0xa7, 0xb3, 0x7, 0x4, 0x31, 0xe6, 0xa8, 0xda, 0x8c}}
 	return a, nil
 }
@@ -239,7 +239,7 @@ func reissuance_reminderTxt() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "reissuance_reminder.txt", size: 1705, mode: os.FileMode(0644), modTime: time.Unix(1659970605, 0)}
+	info := bindataFileInfo{name: "reissuance_reminder.txt", size: 1705, mode: os.FileMode(0644), modTime: time.Unix(1663788021, 0)}
 	a := &asset{bytes: bytes, info: info, digest: [32]uint8{0x91, 0x7f, 0x22, 0xec, 0x13, 0x88, 0xf8, 0xe9, 0x95, 0xd9, 0xb7, 0xc3, 0x49, 0xec, 0x9a, 0x9c, 0x68, 0xb0, 0xc3, 0xfb, 0xa1, 0x36, 0xe3, 0x11, 0x9a, 0x56, 0x43, 0x1f, 0x97, 0xdc, 0x10, 0xfd}}
 	return a, nil
 }
@@ -259,7 +259,7 @@ func reissuance_startedHtml() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "reissuance_started.html", size: 1280, mode: os.FileMode(0644), modTime: time.Unix(1659970605, 0)}
+	info := bindataFileInfo{name: "reissuance_started.html", size: 1280, mode: os.FileMode(0644), modTime: time.Unix(1663788021, 0)}
 	a := &asset{bytes: bytes, info: info, digest: [32]uint8{0x85, 0x1c, 0x4d, 0x46, 0x7e, 0xd3, 0xde, 0xfd, 0xfd, 0x91, 0xc9, 0xf2, 0xf5, 0x4e, 0x7e, 0x7b, 0x63, 0xd5, 0xb, 0xfb, 0x26, 0xdb, 0xb7, 0xcc, 0xa1, 0xf7, 0x72, 0xd, 0x62, 0x8c, 0x7d, 0x6d}}
 	return a, nil
 }
@@ -279,7 +279,7 @@ func reissuance_startedTxt() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "reissuance_started.txt", size: 981, mode: os.FileMode(0644), modTime: time.Unix(1659970605, 0)}
+	info := bindataFileInfo{name: "reissuance_started.txt", size: 981, mode: os.FileMode(0644), modTime: time.Unix(1663788021, 0)}
 	a := &asset{bytes: bytes, info: info, digest: [32]uint8{0x8f, 0xfc, 0x71, 0xd2, 0x64, 0x22, 0xbc, 0x55, 0x7f, 0xb4, 0x9b, 0xaf, 0xfe, 0x3d, 0x12, 0xc3, 0xe8, 0xb1, 0x9c, 0x11, 0x1e, 0xdd, 0x3, 0x1e, 0x88, 0x9, 0xbd, 0x3f, 0x4e, 0xce, 0x9e, 0x25}}
 	return a, nil
 }
@@ -299,7 +299,7 @@ func reject_registrationHtml() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "reject_registration.html", size: 601, mode: os.FileMode(0644), modTime: time.Unix(1659970605, 0)}
+	info := bindataFileInfo{name: "reject_registration.html", size: 601, mode: os.FileMode(0644), modTime: time.Unix(1663788021, 0)}
 	a := &asset{bytes: bytes, info: info, digest: [32]uint8{0xd1, 0xb6, 0x61, 0xf7, 0x52, 0x2d, 0x42, 0xea, 0x17, 0x95, 0x46, 0xb5, 0x77, 0x3, 0xb8, 0x4c, 0x5d, 0xce, 0xdb, 0xd, 0x4, 0xe1, 0x0, 0x5, 0xd2, 0xc4, 0xcc, 0x16, 0xad, 0x58, 0xb6, 0x7d}}
 	return a, nil
 }
@@ -319,7 +319,7 @@ func reject_registrationTxt() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "reject_registration.txt", size: 476, mode: os.FileMode(0644), modTime: time.Unix(1659970605, 0)}
+	info := bindataFileInfo{name: "reject_registration.txt", size: 476, mode: os.FileMode(0644), modTime: time.Unix(1663788021, 0)}
 	a := &asset{bytes: bytes, info: info, digest: [32]uint8{0x2c, 0xda, 0xd2, 0xcb, 0x49, 0xf2, 0x7a, 0x83, 0x7b, 0x67, 0x37, 0xee, 0x2d, 0xd7, 0x4f, 0xbe, 0x1a, 0x66, 0xb0, 0xea, 0x7, 0xc6, 0x87, 0x8, 0xe5, 0xed, 0x39, 0x2c, 0x79, 0x46, 0x26, 0x6a}}
 	return a, nil
 }
@@ -339,7 +339,7 @@ func review_requestHtml() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "review_request.html", size: 1225, mode: os.FileMode(0644), modTime: time.Unix(1634667864, 0)}
+	info := bindataFileInfo{name: "review_request.html", size: 1225, mode: os.FileMode(0644), modTime: time.Unix(1661780575, 0)}
 	a := &asset{bytes: bytes, info: info, digest: [32]uint8{0x3a, 0xaf, 0xd0, 0xf, 0x60, 0x6c, 0x66, 0x89, 0xd8, 0x30, 0x12, 0x82, 0x14, 0xb6, 0xa2, 0x82, 0xa6, 0x1a, 0x2a, 0x94, 0xca, 0xa2, 0x2b, 0x7a, 0x3e, 0xe8, 0xa6, 0xa1, 0x7a, 0x1a, 0x6a, 0x81}}
 	return a, nil
 }
@@ -359,7 +359,7 @@ func review_requestTxt() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "review_request.txt", size: 978, mode: os.FileMode(0644), modTime: time.Unix(1634667864, 0)}
+	info := bindataFileInfo{name: "review_request.txt", size: 978, mode: os.FileMode(0644), modTime: time.Unix(1661780575, 0)}
 	a := &asset{bytes: bytes, info: info, digest: [32]uint8{0x8a, 0x58, 0xb6, 0xd4, 0x10, 0xd8, 0x3d, 0xe1, 0xce, 0x64, 0x58, 0x97, 0x33, 0xa8, 0xef, 0x45, 0xf5, 0xa7, 0xc7, 0xd5, 0x9f, 0xe4, 0x56, 0xdb, 0xc2, 0xa5, 0xc5, 0xec, 0x21, 0xc1, 0xa3, 0xe5}}
 	return a, nil
 }
@@ -379,7 +379,7 @@ func verify_contactHtml() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "verify_contact.html", size: 666, mode: os.FileMode(0644), modTime: time.Unix(1659970605, 0)}
+	info := bindataFileInfo{name: "verify_contact.html", size: 666, mode: os.FileMode(0644), modTime: time.Unix(1663788021, 0)}
 	a := &asset{bytes: bytes, info: info, digest: [32]uint8{0x5e, 0xc9, 0xed, 0x24, 0x65, 0xb2, 0x8b, 0xa7, 0xb1, 0xcb, 0x2b, 0x6c, 0xc9, 0xf9, 0x29, 0x6d, 0x9c, 0x7f, 0x76, 0xaf, 0x6d, 0xce, 0xe9, 0xf1, 0xbc, 0xbc, 0xe4, 0x4a, 0x6b, 0xe9, 0x1, 0xcf}}
 	return a, nil
 }
@@ -399,7 +399,7 @@ func verify_contactTxt() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "verify_contact.txt", size: 536, mode: os.FileMode(0644), modTime: time.Unix(1659970605, 0)}
+	info := bindataFileInfo{name: "verify_contact.txt", size: 536, mode: os.FileMode(0644), modTime: time.Unix(1663788021, 0)}
 	a := &asset{bytes: bytes, info: info, digest: [32]uint8{0xfc, 0x55, 0x37, 0x5e, 0x1, 0xc5, 0x79, 0xd0, 0x2b, 0xbd, 0x1e, 0x7f, 0x8e, 0xa3, 0x8a, 0x8c, 0xe9, 0x7d, 0x5c, 0xcf, 0xff, 0x2f, 0x17, 0xac, 0xf7, 0x18, 0xc8, 0xa3, 0xff, 0x18, 0x86, 0x8a}}
 	return a, nil
 }
